Made code 64-bit clean.  Enabled use of 64-bit arithmetic for integers.
git-svn-id: file:///svn/vx-scheme-20080617-repo/trunk@13 bee25f81-8ba7-4b93-944d-dfac3d1a11cc

+++ src/io.cpp
@@ -233,7 +233,7 @@ TOP:
 	    // hex constant.  Drop the 'x' and convert with strtoul.
 
 	    char * endptr;
-	    unsigned long ul = strtoul (lexeme.str () + 1, &endptr, 16);
+	    uintptr_t ul = strtoul (lexeme.str () + 1, &endptr, 16);
 	    
 	    if (*endptr == '\0')
 		READ_RETURN (make_int (ul));
@@ -425,7 +425,7 @@ void Cell::write (sstring& ss) const {
     switch(t) { 
       case Int: { 
         char buf[40];
-        sprintf(buf, "%d", IntValue());
+        sprintf(buf, "%" PRIdPTR, IntValue());
         ss.append(buf);
         break;
       }
